requests: align with new requests states
- Align code with refactored states and permissions in invenio-requests.

- Adds a new need "CommunityRoleNeed".

- Removed unused code.
