+++ invenio_communities/communities/resolver.py
@@ -17,22 +17,21 @@ entry point.
 from invenio_records_resources.references.resolvers.records import \
     RecordPKProxy, RecordResolver
 
+from ..generators import CommunityRoleNeed
 from .records.api import Community
-from .services.permissions import CommunityNeed
 
 
 class CommunityPKProxy(RecordPKProxy):
     """Resolver proxy for a Record entity using the UUID."""
 
-    def get_need(self):
-        """Return the user need of the community's owner."""
-        # TODO this may become difficult once there's multiple levels
-        #      of membership (owner, manager, curator, ...)
-        #      -> which needs should be generated? None, and let the
-        #         user create the set of required needs from the
-        #         resolved entity? or keep the 'owner' need?
+    def get_needs(self, ctx=None):
+        """Return community member need."""
+        ctx = ctx or {}
+        roles = ctx.get(
+            'community_roles', ['owner', 'manager', 'curator', 'reader']
+        )
         comid = str(self._parse_ref_dict_id(self._ref_dict))
-        return CommunityNeed(comid)
+        return [CommunityRoleNeed(comid, role) for role in roles]
 
 
 class CommunityResolver(RecordResolver):
