+++ invenio_communities/communities/services/__init__.py
@@ -10,11 +10,9 @@
 
 from .config import CommunityFileServiceConfig, CommunityServiceConfig, \
     SearchOptions
-from .permissions import CommunityNeed
 from .service import CommunityService
 
 __all__ = (
-    'CommunityNeed',
     'CommunityService',
     'CommunityServiceConfig',
     'CommunityFileServiceConfig',
