+++ invenio_communities/config.py
@@ -9,38 +9,7 @@
 """Default configuration."""
 
 from invenio_communities.communities.services import facets
-
-COMMUNITIES_MEMBERSHIP_REQUESTS_CONFIRMLINK_EXPIRES_IN = 1000000
-
-SUPPORT_EMAIL = 'test@email.org'
-
-COMMUNITIES_RECORD_INDEX = 'records'
-
-# TODO: Use vocabulary-based model
-COMMUNITIES_DOMAINS = [
-    {'text': 'Agricultural and Veterinary Sciences', 'value': 'agricultural_and_veterinary_sciences'},
-    {'text': 'Biological Sciences', 'value': 'biological_sciences'},
-    {'text': 'Built Environment and Design', 'value': 'built_environment_and_design'},
-    {'text': 'Chemical Sciences', 'value': 'chemical_sciences'},
-    {'text': 'Commerce, Management, Tourism and Services', 'value': 'commerce_management_tourism_and_services'},
-    {'text': 'Earth Sciences', 'value': 'earth_sciences'},
-    {'text': 'Environmental Sciences', 'value': 'environmental_sciences'},
-    {'text': 'Economics', 'value': 'economics'},
-    {'text': 'Education', 'value': 'education'},
-    {'text': 'Engineering', 'value': 'engineering'},
-    {'text': 'Technology', 'value': 'technology'},
-    {'text': 'History and Archaeology', 'value': 'history_and_archaeology'},
-    {'text': 'Information and Computing Sciences', 'value': 'information_and_computing_sciences'},
-    {'text': 'Language, Communication and Culture', 'value': 'language_communication_and_culture'},
-    {'text': 'Law and Legal Studies', 'value': 'law_and_legal_studies'},
-    {'text': 'Mathematical Sciences', 'value': 'mathematical_sciences'},
-    {'text': 'Medical and Health Sciences', 'value': 'medical_and_health_sciences'},
-    {'text': 'Philosophy and Religious Studies', 'value': 'philosophy_and_religious_studies'},
-    {'text': 'Physical Sciences', 'value': 'physical_sciences'},
-    {'text': 'Psychology and Cognitive Sciences', 'value': 'psychology_and_cognitive_sciences'},
-    {'text': 'Studies in Creative Arts and Writing', 'value': 'studies_in_creative_arts_and_writing'},
-    {'text': 'Studies in Human Society', 'value': 'studies_in_human_society'},
-]
+from flask_babelex import lazy_gettext as _
 
 COMMUNITIES_ROUTES = {
     'frontpage': '/communities',
@@ -70,3 +39,27 @@ COMMUNITIES_FACETS = {
     }
 }
 """Available facets defined for this module."""
+
+
+COMMUNITIES_ROLES = {
+    'reader': dict(
+        title=_('Reader'),
+        description=_('Can view restricted records'),
+    ),
+    'curator': dict(
+        title=_('Curator'),
+        description=_('Can edit and view restricted records'),
+    ),
+    'manager': dict(
+        title=_('Manager'),
+        description=_('Can manage members, curate records and view restricted records'),
+        can_manage_members=['manager', 'curator', 'reader'],
+    ),
+    'owner': dict(
+        title=_('Owner'),
+        description=_('Full administrative access to the entire community.'),
+        can_manage_members=['owner', 'manager', 'curator', 'reader'],
+        is_owner=True
+    ),
+}
+"""Community roles."""
