+++ setup.cfg
@@ -26,7 +26,7 @@ packages = find:
 python_requires = >=3.6
 zip_safe = False
 install_requires =
-    invenio-requests>=0.3.0,<0.4.0
+    invenio-requests>=0.3.1,<0.4.0
     invenio-vocabularies>=0.11.0,<0.12.0
 
 [options.extras_require]
